Project cleanup
- dummy model, services, repos and routes deleted
- removed database migrations scripts and env configurator
- DatabaseConnection class moved to utils.database
