--- src/main/kotlin/mobi/waterdog/rest/template/core/persistance/PartRepository.kt
@@ -1,11 +1,0 @@
-package mobi.waterdog.rest.template.core.persistance
-
-import mobi.waterdog.rest.template.core.model.Part
-
-interface PartRepository {
-    fun list(): List<Part>
-    fun delete(partNo: Long)
-    fun count(): Int
-    fun getPartsForCar(carId: Long): List<Part>
-    fun addPartToCar(carId: Long, part: Part): Part
-}
