--- src/main/kotlin/mobi/waterdog/rest/template/core/persistance/sql/CarMappingsTable.kt
@@ -1,8 +1,0 @@
-package mobi.waterdog.rest.template.core.persistance.sql
-
-import org.jetbrains.exposed.dao.id.LongIdTable
-
-internal object CarMappingsTable : LongIdTable("cars") {
-    val brand = varchar("brand", length = 255)
-    val model = varchar("model", length = 255)
-}
