--- src/main/kotlin/mobi/waterdog/rest/template/core/service/CarServiceImpl.kt
@@ -1,71 +1,0 @@
-package mobi.waterdog.rest.template.core.service
-
-import mobi.waterdog.rest.template.conf.database.DatabaseConnection
-import mobi.waterdog.rest.template.core.exception.AppException
-import mobi.waterdog.rest.template.core.exception.ErrorCode
-import mobi.waterdog.rest.template.core.model.Car
-import mobi.waterdog.rest.template.core.model.CarSaveCommand
-import mobi.waterdog.rest.template.core.model.Part
-import mobi.waterdog.rest.template.core.model.RegisterPartReplacementCommand
-import mobi.waterdog.rest.template.core.persistance.CarRepository
-import mobi.waterdog.rest.template.core.persistance.PartRepository
-import mobi.waterdog.rest.template.core.utils.pagination.PageRequest
-import org.koin.core.KoinComponent
-import org.koin.core.inject
-import org.slf4j.LoggerFactory
-
-class CarServiceImpl : KoinComponent, CarService {
-
-    private val carRepository: CarRepository by inject()
-    private val partRepository: PartRepository by inject()
-    private val dbc: DatabaseConnection by inject()
-
-    companion object {
-        private val log = LoggerFactory.getLogger(this::class.java)
-    }
-
-    override suspend fun count(pageRequest: PageRequest): Int {
-        return dbc.suspendedQuery {
-            log.info("Counting cars from the repository")
-            carRepository.count(pageRequest)
-        }
-    }
-
-    override suspend fun exists(carId: Long): Boolean = dbc.suspendedQuery { carRepository.exists(carId) }
-
-    override suspend fun getCarById(carId: Long): Car? {
-        return dbc.suspendedQuery { carRepository.getById(carId) }
-    }
-
-    override suspend fun insertNewCar(newCar: CarSaveCommand): Car {
-        return dbc.suspendedQuery { carRepository.save(newCar) }
-    }
-
-    override suspend fun updateCar(car: Car): Car {
-        return dbc.suspendedQuery {
-            if (!exists(car.id)) {
-                throw AppException(ErrorCode.NotFound, "Could not find car with id '${car.id}'.")
-            }
-            carRepository.update(car)
-        }
-    }
-
-    override suspend fun registerPartReplacement(replacedParts: RegisterPartReplacementCommand): Car {
-        // this runs the operation as a single transaction
-        return dbc.suspendedQuery {
-            val car = carRepository.getById(replacedParts.carId)
-            requireNotNull(car) { "Car must exist" }
-            for (part: Part in replacedParts.parts) {
-                partRepository.addPartToCar(car.id, part)
-            }
-
-            car
-        }
-    }
-
-    override suspend fun list(pageRequest: PageRequest): List<Car> {
-        return dbc.suspendedQuery {
-            carRepository.list(pageRequest)
-        }
-    }
-}
