--- src/main/kotlin/mobi/waterdog/rest/template/core/service/PersonServiceImpl.kt
@@ -1,38 +1,0 @@
-package mobi.waterdog.rest.template.core.service
-
-import mobi.waterdog.rest.template.conf.database.DatabaseConnection
-import mobi.waterdog.rest.template.core.model.Person
-import mobi.waterdog.rest.template.core.persistance.PersonRepository
-import mobi.waterdog.rest.template.core.utils.pagination.PageRequest
-import org.koin.core.KoinComponent
-import org.koin.core.inject
-
-class PersonServiceImpl : KoinComponent, PersonService {
-
-    private val personRepository: PersonRepository by inject()
-    private val dbc: DatabaseConnection by inject()
-
-    override suspend fun add(person: Person): Person {
-        return dbc.suspendedQuery { personRepository.save(person) }
-    }
-
-    override suspend fun update(person: Person): Person {
-        return dbc.suspendedQuery { personRepository.update(person) }
-    }
-
-    override suspend fun getById(personId: Int): Person? {
-        return dbc.suspendedQuery { personRepository.getById(personId) }
-    }
-
-    override suspend fun deleteById(personId: Int) {
-        dbc.suspendedQuery { personRepository.delete(personId) }
-    }
-
-    override suspend fun count(pageRequest: PageRequest): Int {
-        return dbc.suspendedQuery { personRepository.count(pageRequest) }
-    }
-
-    override suspend fun list(pageRequest: PageRequest): List<Person> {
-        return dbc.suspendedQuery { personRepository.list(pageRequest) }
-    }
-}
