+++ src/main/kotlin/mobi/waterdog/rest/template/core/utils/exception/ExceptionHandler.kt
@@ -1,4 +1,4 @@
-package mobi.waterdog.rest.template.core.exception
+package mobi.waterdog.rest.template.core.utils.exception
 
 import io.ktor.application.call
 import io.ktor.features.StatusPages
