--- src/main/kotlin/mobi/waterdog/rest/template/core/utils/healthcheck/HealthCheck.kt
@@ -1,28 +1,0 @@
-package mobi.waterdog.rest.template.core.utils.healthcheck
-
-import kotlinx.coroutines.withTimeout
-import mobi.waterdog.rest.template.conf.database.DatabaseConnection
-import org.koin.core.KoinComponent
-import org.koin.core.inject
-
-internal class HealthCheckInjector : KoinComponent {
-    val dbc: DatabaseConnection by inject()
-}
-
-fun Health.Configuration.liveness(timeoutMs: Long) {
-    readyCheck("database") { checkDatabase(timeoutMs) }
-    liveCheck("alive") { true }
-}
-
-fun Health.Configuration.readiness(timeoutMs: Long) {
-    readyCheck("database") { checkDatabase(timeoutMs) }
-}
-
-private suspend fun checkDatabase(timeout: Long): Boolean =
-    try {
-        withTimeout(timeout) {
-            HealthCheckInjector().dbc.ping()
-        }
-    } catch (ex: Exception) {
-        false
-    }
