+++ src/main/resources/application.conf
@@ -10,25 +10,3 @@ ktor {
     modules = [mobi.waterdog.rest.template.ApplicationKt.module]
   }
 }
-
-datasource {
-  driver = org.postgresql.Driver
-  jdbcUrl = ${?JDBC_URL}
-  username = ${?DB_USERNAME}
-  password = ${?DB_PASSWORD}
-  pool {
-    defaultIsolation = "TRANSACTION_REPEATABLE_READ"
-    defaultIsolation = ${?DB_CONN_DEFAULT_ISOLATION}
-    maxPoolSize = 5
-    maxPoolSize = ${?DB_CONN_POOL_SIZE}
-  }
-}
-
-healthcheck {
-  readiness {
-    database {
-      timeoutMillis = 3000
-      timeoutMillis = ${?LIVENESS_TIMEOUT_MILLIS}
-    }
-  }
-}
