--- src/main/resources/db/migration/V1.2__persons.sql
@@ -1,6 +1,0 @@
-CREATE TABLE persons
-(
-    id       BIGSERIAL PRIMARY KEY,
-    name     VARCHAR(255) NOT NULL,
-    birthday VARCHAR(255) NOT NULL
-);
