--- src/test/kotlin/mobi/waterdog/rest/template/core/model/TestModels.kt
@@ -1,49 +1,0 @@
-@file:UseSerializers(InstantAsStringSerializer::class)
-
-package mobi.waterdog.rest.template.core.model
-
-import kotlinx.serialization.KSerializer
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.UseSerializers
-import kotlinx.serialization.descriptors.PrimitiveKind
-import kotlinx.serialization.descriptors.PrimitiveSerialDescriptor
-import kotlinx.serialization.descriptors.SerialDescriptor
-import kotlinx.serialization.encoding.Decoder
-import kotlinx.serialization.encoding.Encoder
-import java.time.Instant
-import java.time.format.DateTimeFormatter
-
-@Serializable
-data class TestInstantStringSerialization(val date: Instant)
-
-object InstantAsStringSerializer : KSerializer<Instant> {
-    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("Instant", PrimitiveKind.STRING)
-    override fun serialize(encoder: Encoder, value: Instant) =
-        encoder.encodeString(DateTimeFormatter.ISO_INSTANT.format(value))
-
-    override fun deserialize(decoder: Decoder): Instant = Instant.parse(decoder.decodeString())
-}
-
-@Serializable
-data class TestInstantLongSerialization(@Serializable(with = InstantAsLongSerializer::class) val date: Instant)
-
-object InstantAsLongSerializer : KSerializer<Instant> {
-    override val descriptor: SerialDescriptor = PrimitiveSerialDescriptor("Instant", PrimitiveKind.LONG)
-    override fun serialize(encoder: Encoder, value: Instant) = encoder.encodeLong(value.toEpochMilli())
-    override fun deserialize(decoder: Decoder): Instant = Instant.ofEpochMilli(decoder.decodeLong())
-}
-
-@Serializable
-class TestSealedClass(val exp: Expr)
-
-@Serializable
-sealed class Expr
-
-@Serializable
-class Const(val number: Double) : Expr()
-
-@Serializable
-data class Sum(val e1: Expr, val e2: Expr) : Expr()
-
-@Serializable
-object NotANumber : Expr()
